\name{dot-PBSadmbEnv}
\docType{data}
\alias{.PBSadmbEnv}
\title{PBSadmb Environment}
\description{
  An environment set aside for PBSadmb.
}
\usage{.PBSadmbEnv}
\format{
A new environment with a \code{.GlobalEnv} parent.
}
\source{
  Generated by a call to the base function \code{new.env()}.
}
\details{
  The environment is created in \code{'zzz.r'} and is used by
  \code{PBSadmb} functions \code{'alisp'}, \code{'atget'}, 
  \code{'atput'}, \code{'atprint'}, and \code{'atcall'}.
}
\seealso{
In \pkg{PBSadmb}:\cr
  \code{\link[PBSadmb]{alisp}},
  \code{\link[PBSadmb]{atget}}\cr
In \pkg{PBSmodelling}:\cr
  \code{lisp},
  \code{tget}
}
\keyword{environment}

