\name{installADMB}
\alias{installADMB}
\title{ Install windows admb binary (for gcc) }
\description{
  Only applicable for Windows: Downloads and installs the windows ADMB binary for gcc.
  ADMB is installed under PBSadmb's library directory under R.
}
\usage{
installADMB()
}
\value{
	The path where ADMB was installed.
}
\keyword{file}
